refactor: add support for kubeconfig command
