+++ internal/cli/cli.go
@@ -18,6 +18,8 @@ func Run(ctx context.Context) error {
 		Long:    "Nais platform utility cli, respects consoledonottrack.com",
 		Version: version + "-" + commit,
 	}
+	fs := app.PersistentFlags()
+	fs.BoolP("verbose", "v", false, "Verbose output")
 
 	app.AddCommand(
 		login(),
